backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "jah0m", email = "jahom990831@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "openai (>=1.75.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

